pyproject.toml
[project]
name = "tubesiphon"
version = "0.1.0"
description = "Python CLI for fetching YouTube subtitles into local files."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
    "yt-dlp>=2026.3.17",
]

[project.scripts]
tube-siphon = "tubesiphon.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
